server/requirements.txt
altgraph==0.17.4
asgiref==3.8.1
blinker==1.9.0
certifi==2024.8.30
click==8.1.7
cloudinary==1.41.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==5.1.7
djangorestframework==3.16.0
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-PyMongo==3.0.1
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
graphviz==0.20.3
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pymongo==4.11.3
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pywin32-ctypes==0.2.3
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
Werkzeug==3.1.3
